Incorrect Math Caused by CINT vs INT
Sorry about the script freeze, but it came to our attention that the
math in the expedited screening was off because of the use of CINT. CINT
rounds to the nearest integer whereas INT does not. A client with
$149.99 in income was being considered not expedited (hypothetical
scenario) because the income was being rounded to $150. Similarly, a
client with $100.25 in liquid assets was being considered expedited
because CINT was rounding it DOWN to $100.
We changed CINT to INT on income, assets, and rent as the actual amount
should be used in these situations and not a rounded amount (utilities
was left alone b/c those are whole number deductions and not actual
amounts.