Added logic to check first paydate against others
Basically, July 4th was a holiday. It was also a Friday, so some folks
got paid. Other folks got paid on the Thursday before. Unfortunately,
the script saw that, thought the paydates were Thursday, and incorrectly
prospected the income accordingly.

The logic I just installed will catch if all paydates are not on the
same day of the week (for biweekly and weekly income only), and end the
script. This could be thought of as a temporary workaround until we can
build better logic to handle these situations.