Income start date can be the future
The script was incorrectly declaring that the income start date needed
to be the current date or prior. I've fixed the logic by removing the
lines that required this (it was the first part of an If...elseif...else
statement). I also added the MAXIS_case_number_finder function while I
was all up in there.