Create NOTE - ACA Renewal
A script that modifies the notices for HC renewals for apparently magi populations, when the renewal is not received but HC is approved per new rules and guidelines.  (and case notes the action).   Currently relies on worker input to prevent the modification of notices for magi members in mixed households, would like to add a check that only modifies approval notices.