MainForm.vb update (downloads files from GitHub)
I'm not ready for the rest of this to go yet, but this is one hell of an
update, and I want to commit it to be able to revert back here in the
event of an issue.

This update will dynamically grab the directory from GitHub and strip
out all but the script files folder. This will mean folks will not have
to go into SIR to get scripts, and we can push out updates just by
instructing users to "reclick". Committing now so as not to lose this
work. :)