MNS Retro in MAXIS
Alrighty everyone! I need some fresh eyes on this problem. Everything
should be working on these with the exception of the "Quick Fail" Person
test function. The issue that I am having is that when when run. It is
only doing 1 month before it exits the loop. What's odd is that when run
by itself it seems to have no issues. I have used MsgBox's to ensure
that the variables have the correct values and they do. Could a fresh
set of eyes perhaps take a look and let me know if I'm missing
something! The for loop before it works just fine so I don't think its
the loop or missing syntax or something and when it is run outside the
script it has no issues... I'm feeling a little lost on this one!