Removed config folder
Basically, I don't want this mucking up the beta counties who are
choosing to download this work. Plus, this will work better with the new
config program I'm finishing. My repo for the new version is
https://github.com/theVKC/BlueZone-Scripts-Installer, and I encourage
folks to download/mod/update. I can move that repo to the DHS (or
county) groups if it ever comes up. :)