MNS Source full upload
This is a full upload of all my current working source. I have
"projects" but they are more theory crafting at this point. This is the
installer that I have and the BZ patch that I made. The patch basically
copies a BlueZone profile from a dir to another dir for people that have
problems with their settings "sticking". Not sure if this is only a
problem at SLC but I love collaboration so if you need it here it is
:-). Veronica if you want to look at the installer go for it but I was
wondering what you thought about just adding an "add MNSure Scripts"
check box to your current installer for those that want to try these
out. It's up to you, if not my installer copies over into the dir they
install the current scripts to and patches in the MNSure info. I have to
update the scripts to be in an installable format so please be aware
that I will create a zipped installer for release when we figure out how
we want to move forward. I'm looking over the GPL right now and I'll
send an email out today :-)