Client contact updates
The "contact type" dropdown was replaced with a combobox. Also, EZ info
references were removed and replaced with "email". Since it's a combobox
folks can now add whatever they like to this.

I also updated the logic to make it required, and fixed the tabbing of
the script while I was in there.