HC Retro Fiat Script
Hey guys here is the HC Retro Fiat script. It's going to need some
tweaks of course but this should be ok to get started. Try it out. I use
case number 209700(or 207900?) in the training region for testing my
scripts. Check it out and let me know what you think.