New function: worker_county_code_determination
This function determines the county code for agencies with more than one
county. So, for example, if DHS staff are using the script, or if
Southwest Human Services is using a script across county lines. It
prompts the user to enter a county code "to proxy as", if the installer
sets them up with a worker_county_code of "MULTICOUNTY". I made that
change over on my repo for the installer. :)

I've updated all sups nav scripts, as well as FUNCTIONS FILE and BULK -
pull data into excel. I also updated the Appointment Letter script to
remove the word "County" from the memo (as this is now going to be added
by the installer).