Fixed issue with MONT navigation
If a worker was not working their own caseload (ie case banking), and case noting for MONT closures, the script was putting the x1 number in the wrong column. Basically, it was reading "FAM" from the dialog, but entering the entire x1 number where just the last three digits needed to go. This resulted in "x1##x1#" being entered into MONT.

I corrected the column, it should now work properly.