Create SNAP E&T Orientation Letter
Putting this up in case anyone wants to modify for their county.  This script is very Saint Louis County specific currently.  We have 3 separate offices providing SNAP E&T orientations, so their information is entered directly into the script.  If this were to be widely deployed it would mean a new array of addresses / phone numbers for the config utility and functions file, as well as adding a user input for appointment times.