Added Andrew Fink's code for "penny issue"
This enhancement came courtesy of Andrew Fink of Dakota County (former
FW, now moved on from the department, apparently). A while back he
actually solved the bug listed in SIR as a non-critical called "Dail
Scrubber- Child Support". His fix is below. There aren't a ton of
comments in the code, but I tested it on a few fake CSES (TIKL) messages
and it seemed to work well. Essentially, it rounds down, sticks the
remainder in a cell, and then adds that remainder back onto the first HH
member. Pretty slick.

I will be revamping this script in the coming weeks with more comments,
in anticipation of DHS staff deciding a new definition of "within the
realm". So, I didn't want to get too bogged down in adapting this. Go
Dakota!