Moved MFIP orientation to alpha scripts
No worries, @courtrightd , it's not because I don't love it. :)

Some agencies have their WF1 (and orientation locations) in a separate
office from their primary office. In order to get this tested in
multiple agencies, I'll first need to add a field to the installer. I
can do that, but not this morning. I have a bug fix that needs testing
ASAP, so I'm moving this to alpha (so as not to send to other agencies).

I'll be doing a bit of tweaking to the script before it goes live, to
make the case note look similar to the others (bullet points, etc), and
to tweak the MEMO to use the new write_new_line_in_MEMO function. Minor
tweaks. Solid, solid script though.

(Oh! And I added a file extension (.vbs) to the file.)