Updated CS-hire script with bug fixes and comments
Tons o'changes:

1. Created a new dialog for "job known to agency", as this script's
dialog appears to have been lost.
2. Modified the scripts dialog to have the worker sig on the first
dialog. Now, there's only one dialog. :) Also, the "is the job known"
dropdown box is replaced with a checkbox.
3. A bug was reported which was causing the "job known" if...then to
fail if "yes" was selected (it would add panels anyway). That was due to
the dialog being called "job_known" and the dropdownlist variable being
called "job_known". The variable could never equal "yes" because it
equaled a dialog. So, I changed the name of both, and adjusted variable
references throughout the script. Bug solved.
4. Added comments throughout, as my commenting was hasty back when I
first made this script.
5. Lots of lines of this have since been superceded by FUNCTIONS FILE
changes. I added a few script_end_procedure references, as well as a
create_MAXIS_friendly_date, and removed the old code.
6. Moved some things around for clarity sake. Expanded an If...then to
be multiline, just because they look nice (71-75).
7. Script will no longer put MM/01/YY in for the income start date. That
isn't necessarily good policy (to guess on that date).

Also, and unrelated: I noticed an old "x102 dialog" in the dialog
folder. Removed it as it's old. #keepacleanshop