Added changed TIKLing to be a checkbox option
A few changes came with this one:

1. Updated the dialog to make TIKLing an optional event, by adding a
checkbox. While I was up-in-there, I also changed the "is job already
known to agency" to be a checkbox (so that it won't default to "Yes"
anymore).
2. Updated script to reflect dialog above.
3. Set TIKL_check to default to "checked".
4. Updated references to job_known variable, to be job_known_check, and
removed "Yes" (replaced with "checked").
5. If create_JOBS_check is checked, it will now case note that the JOBS
panel was updated.
6. If TIKL_check is checked, it will case note that the TIKL was sent.
If it's unchecked, it won't case note it.
7. Script ends right after case note, if the TIKL_check is unchecked.
8. A pet peeve that I never fixed until now: now the TIKL says "if not
received and processED", note emphasis on previously missing -ed suffix.