Fourth major change to PDTE
Split PDTE into subscripts ala DAIL scrubber? Why? Because it had become
so long it was challenging to dissect where I was going. This will be
easier to manage, plus it will allow for the insertion of more dialogs
if we need to.

Now, when I add LTC/GRH list gen to this script, it can be as a subset
as well. Saving button space, team!