Added a dropdown box
The dropdown box allows workers to select the type of approval they are
doing. So they can either select initial, renewal, recertification,
change, reinstate, or enter their own type if none of these fit. It will
then add it to the header of the case note making the case notes more
readable and obvious what kind of approval this was.