Fifth major change to PDTE
REVW and REVS are now live. They create an Excel list, relatively
simply. They will indicate to the user percentages of cases with N, I,
and U status. REVS also asks if you want to go to the current month + 2,
as it's one of the rare few screens that can do that.