Seventh (and final) major change to PDTE
About time I wrapped this up.

1. LTC-GRH list generator is now live. This one creates a list of FACIs,
AREPs, and waiver types for a specified worker. Should be helpful in
LTC/GRH circles, especially around sending verifications and caseload
tracking.
2. There was an old dialog from my Anoka days called "FACI list
generator". I've replaced it with the LTC-GRH version, and enhanced it a
bit with info about multiple X1 numbers.
3. REPT scanning dialog got overhauled to add this feature. I also split
ARST from the other REPTs, since those REPTs are about cases and this
one is about stats.
4. Updated BULK - pull data into Excel to reflect the above.

With that, I'm out of this one for a while. Hopefully it goes well in
the counties. Hopefully. :)