Emergency note changes
Simple changes here:

1. Fixed footer month being blank at times. Also corrected for an issue
where the footer month grabs a $ sign if you're on the HEST screen when
you run it.
2. Replaced references to "thirty_percent_rule_notes" with
"percent_rule_notes", as not all agencies use the same verbiage.
3. Added the new emer_percent_rule_amt variable from the functions file
in. Now, agencies which use a lower or higher amount will have their
amounts show up here.
4. Removed phantom "FSPR_check" logic in the case note. No idea why that
was there, must have been a carry over from HRF or something.
5. Added the crisis info to the case note header. Replaces a period with
a null, because we don't want a period in that header. It'd look weird.