Big 'ol change to Power Pad
Responding to feedback that the Power Pad was "too hard to read", and
facing the reality of adding 2 new scripts and not having a good place
to put them, I've reconfigured the Power Pad to group scripts using
GroupBoxes instead of just a line of scripts. This results in a slightly
larger Power Pad, but with much more space and better grouping. By
popular demand, LTC has been moved to their own "notes" section as well.