Installed fix for FUNCTIONS FILE erroring out on INSA
Basically, what would happen is that if the INSA panel had two spaces in
the Insurance Company name (so something like "BCBS<space><space>MN"),
it'd error out because of the for/next array processing. You can't do a
"right" with a string of zero length.

The fix is a simple If/then, which will ignore words that trim out to be
"" (double spaces).

Also, the semicolon wasn't showing up between INSA and MEDI (which both
use the INSA variable). This is because the autofill function chops off
the ending semicolon, which is needed for MEDI cases with attached INSA.
I inserted a line into the HCAPP script to fix this by adding a
semicolon when there's already text. Later in the script, it was already
clearing that semicolon, so it doesn't matter if it ends up in there at
that point.

Committing to beta, will pull into master once Anoka's BILS updater
fixes come through.