Initial commit for Lisa McNally's CEI premium script
This is a cool one. It's a custom mod of Anoka's old MEDI-CEI script.
This allows for a few more items, to let it be used for non-medicare
CEI. Assuming it works, I'll make it available for the beta counties
ASAP.

I'm running into a dialog is out of range issue and I don't have time to
resolve right now. I've modified Lisa's code to use an array, which can
allow a theoretically endless stream of case numbers to go into this
script. Hopefully, I can get these issues resolved and the script
updated ASAP.