Removed excess dialogs
In order to make this script run in Anoka County, the script had to be
modified. The cancel dialog (to confirm canceling the script) and the
case note dialog (to confirm the case note) have been replaced with
CancelDialog = Msgbox ("...", vbYesNo) and CaseNoteDialog =
MsgBox("...", vbYesNo).

The script no longer runs in to the dialog ceiling.