Update to NOTE - Combined AR
Changing the name of the variable helps the script interact with the
functions file.

It just so happened that I had an MSA renewal with no other programs up
for renewal this month. The script failed to pull through the UNEA
information. Upon closer inspection, "MSA_check_ doesn't appear anywhere
in the Functions File. To get this script to work with the Functions
File, "MSA_check" has been changed to "cash_check"