Added a sort of "debugger mode"
So I have a few "enhancements" I need to make to this script, and it's
tough because it doesn't really allow for easy use in inquiry mode for
debugging. So, I came up with a solution: it'll check for my logon
credentials, and ask me if I want to bypass the date-dependent parts. If
those date-dependent checks are bypassed, it'll just do a MsgBox with
the case note info, and not the actual case note (when I'm testing with
production data, I don't want to generate these case notes). Woohoo!