new case transfer & approved progs
The case transfer script now allows for county-to-county transfers. The
script sends a memo to the new address and the worker has the option of
informing the client the case will close by closure_date if required
verifs are not received.

The approved programs script now autofills cash results just like it has
been doing for SNAP.