Transfer case
Changed Change report form to be just a check rather than having the
client enter the date it was sent. Also changed the way the closure date
works. Before we had a button that entered the last day of the next
available month. However it was causing problems when people didn't send
out change report forms and thus didn't need a closure date on the memo.
As such we changed it to a check box and made the checkbox trigger a
calculation to determine the next available end date for the worker
rather than relying on the worker.