Added @ramonalisa1's CEI premium script
I made some slight modifications from @ramonalisa1's original design:

1. Added an array to simplify the code
2. Modified dialog to allow multiple dialogs' worth of info to be added
to the array
3. A few pieces of functional streamlining based on work we did between
when the original Anoka MEDI CEI script was written, and now

In addition to the script, I modified my workspace and the Power Pad to
show the script.

Give it a try, team! Let's see how it works! Perhaps we can run it
through the ringer after Project Krabappel is through.