A utility for figuring out what "EMWaitReady" is best
This is actually alarming me that I didn't think of this sooner.

For years, I've always just used `EMWaitReady 0, 0`, because I found it
to be stable and reliable over `EMWaitReady 1, 1` (which I used in
earlier scripts). However, I'm beginning to question that assessment.

We may want to run our own speed tests independently, and possibly some
other "stress test" type of activities, but, I figured I'd get the ball
rolling with this simple utility to determine best speed. From here, if
anyone has other ideas for testing this theory, be my guest. :)