REPT Search mods
@robertkalb did a good job with this one, but I made some minor
improvements:

1. Used Notepad++'s awesome column editing to replace the double space
nesting with tab nesting. Just being a geek.
2. Added a MsgBox warning the user that they must enter a seven digit
worker number, as it was just looping without telling the worker why the
validation didn't work.
3. Other than REPT/ACTV, all other scripts showed "REPT/INAC Search
Complete" in the closing box. I modified the closing boxes to match the
screen requested.
4. Replaced `script_end_procedure` with `MsgBox`, and added
`script_end_procedure("")` to the end. The stats collection will show
all of those `script_end_procedure` messages as "errors", so this is
easier.