Major logic overhaul of REVW-MONT closures
The REVW part has some changes which hopefully bring it in line with
CM0009.06.03. This was reported to me on SIR a while back, and is
potentially more error prone of an issue than I anticipated. Basically,
here's the change:

Let's say a client fails to turn in a CAF for 12/14 SNAP recert. Do they
get to turn it in anytime during the month of December? Yes. Is it
considered a new application? **Yes**. Unfortunately, the script's
wording made it "sound like" a REIN (even though it didn't say REIN
anywhere on it). In order to avoid confusion, I've modified the script
to now provide three separate intake dates and docs needed dates:
separate dates for cash, SNAP, and HC.

In addition to all of this, I'm preemptively removing WB (as it's going
away and I'm betting a lot of folks will close generating weird case
notes with this script). The MONT logic has been left unchanged.

Hope this works, it should resolve issue #50.