Removed excel_read and excel_write and updated functions
* excel_read and excel_write were struggling because it appears that you
can't pass the objExcel and objWorkbook to the function (when they're
declared in another function). I made the decision to remove these, but
will continue with the very helpful excel_open function.
* Updated the DIET function to have clearer variable names.
* Updated PBEN to include pben_type, looks like it was missing from the
original.
* Updated PDED and STWK with capital panel variable names, and spaces
between commas (using a CSV to create a dynamic list, and needed the
space for standardization purposes).
* Updated STWK with a clearer refused_employment variable name.