Welcome to the jungle, INAC scrubber
INAC scrubber has been "DHS-ified". For non Anoka folks, it's a brand
new script. For Anoka folks, here's what's different:

* Arrays, arrays galore. All of the work is now done with an array,
which should result in a slight net performance improvement. It's also
much more efficient, and will work in agencies without Microsoft Excel
(although it does require Microsoft Word for the claims document). This
array is 8 dimensional, and the PMI numbers for MMIS are in a sub-array
(that's so complicated, just go look at the script for more info).
* Developer mode is available by entering "UUDDLRLRBA" as the worker
signature. This will disable case noting and XFERing, and will allow it
to work in inquiry. This will also enable a visible Excel spreadsheet of
the main array.
* The case note has been significantly expanded, and now allows for
agencies without a special "CLS" number.
* That "CLS" number is now called from FUNCTIONS FILE.
* Lots of for...next replacing do...loops.

For those of you outside of Anoka, INAC scrubber case notes all of your
INAC cases without MMIS conflicts or existing DAIL messages. It
generates a Word doc with cases with active claims, as well as cases
where ABPS needs to be updated. If your agency specifies a "CLS" account
(in the installer), it'll even XFER the case there. It's a big deal
script in Anoka, and I'm very, very glad to be deploying it (actually,
it was one of the first BULK scripts I ever built, so I feel a little
like it's just come home to its creator).

@CDPotter and @RobertFewins-Kalb, please deploy this in Anoka right
away, so that we can test in Monday when everyone comes in. Glory
awaits!