Made XFER a checkbox, so that APPL'd, PND1, and PND2 cases can be XFERed
This required some fancy tweaks to the dialog, as well as a few
if...thens throughout the script. I also moved the transfer portion to
be a function.

I also changed how it determines what agency the new worker is with.
Before, it was looking at the X1 number, but it has come to my attention
that lots of state-operated agencies (i.e., DHS staff and some tribes)
have "faked out" different county names, in spite of everyone using PW
numbers. So, instead of just comparing the first four digits of your
X1/PW number, it'll navigate to REPT/USER and look up your credentials
that way. I show up as "91", but others might be "90" or "92", or almost
anything else. This was causing issues before in trying to transfer
(script thought we were in county, but MAXIS considers it "out of
county"). Whew. Hopefully that makes sense.