Replaced three Power Pads with one
It's now called "DHS supported". I've added the (coming soon) GitHub
redirect scripts to the pad, as well as links to common websites, and
several new navigation buttons. I kind-of want this to be "fully loaded"
so that way we won't have to do a manual update for a few months. :)