NOTES - APPROVED PROGRAMS
Fix to approved programs.
(1) Removed MsgBox that was still hanging around from development.
(2) Fixed the way the script was reading cash results and case noting.
Previously, it was reading a string of 5 characters and manipulating
them in a convoluted way to convert it to 5 digits. Now the script just
reads a string of 8 characters, replaces " " with "0" and case notes
using FormatCurrency(x) ... although, obviously, I didn't use "x" as a
variable name.