CAF updated
Here's what I modified:

* Added a checkbox for `eDRS_checkbox`. That's the reason for the
season, after all.
* I also took the opportunity and broke apart the checkboxes on dialog
part 3 into two groups: one for "common elements workers should case
note" and another (which was moved to the bottom of the dialog, aligning
with several other scripts) for "actions the script can do". I also
alphabetized the common elements while I was up-in-there.
* Replaced `_check` with `_checkbox` where it represents a checkbox.
This is to align with what I teach people now (think about it, it's more
logical).
* Replaced earlier `MAXIS_check` with the new `check_for_MAXIS(True)`
function.
* Removed ERRR prone checking, as that is now included in
`navigate_to_screen`.
* Replaced instances of `= 1` with `= checked`, to align with how I
teach people.
* Replaced all case note functions with their new equivalents. This
resulted in removing all of the `, 6)` parameters. It looks cleaner now.
I will continue to modify case notes with these new functions as I work
on them, and I ask my fellow scriptwriters to do the same. :)

I'll keep this branch running for possible issues, as well as tackling
the CSR tomorrow. For now, it's 9:20PM and time for bed.