Added Hennepin handling to remove two scripts
Because Hennepin is (at least temporarily) scheduling appointments in a
time-range (instead of a specific time), and they (at least temporarily)
have permission to do this, the NOMI and Appointment Letter scripts
might technically provide incorrect notice to clients, which could
create policy errors. So, I am temporarily removing access to them from
Hennepin users. If Hennepin is to use this same method in the near
future, separate versions of these two scripts may be deployed to allow
for a range of times.