NOTES - DRUG FELON.vbs
This maxis case note would be utilizied for SNAP clients who have been convicted of a drug felony in the last 10 years and have to comply with UA's. This case note would track their UA results/test days and update case if changes need to be made to their benefits due to results of the UA's.