Removed "MAXIS" from navigate_to_screen
For some reason (and I have been chasing this bug all night) this isn't
working. I will explain in a FuncLib issue report. This will require
some testing to determine what (if anything) is wrong. I think that
somewhere, somehow, case_number (the variable) isn't being pulled for
navigate_to_MAXIS_screen, but for some absolutely bizarre reason, is
being pulled for navigate_to_screen, in a way which may be chalked up to
a ghost or something. No clue here, but this seems to fix it...