LTC transfer penalty didn't have a file extension
Easy mistake @IlseFerris. Apparently it was missing the .vbs.

While I was in there, I removed DIMs for the time being to test a bug.
When you write scripts, keep doing DIMing. I think a solution will be
close at hand but I'm commenting these out for now just to provide a
quick fix to the beta counties. You can ask me more about that later if
you so desire. :)