Updated array handling
Hey @IlseFerris! You might be interested in this... the issue was that
county_FSET_offices was not assigned as an array right away. So, the
logic which says `county_FSET_offices = ""` doesn't fly if it is, in
fact, an array. So, I set up a blank array right away, and BOOM! Works.