Fixed a bug and updated some functions
An "End if" was missing from the FuncLib call. It had been incorrectly
placed after the dialog, which would let it work locally but not on
GitHub (as GitHub only loads the FuncLib call once, it was skipping the
part where it actually dialogs).

I fixed that by reinserting the missing end if. I also updated a couple
of functions while I was in there. Nothing major.