committing to pull request: lots of updates
removed old functions, added new functions.  Cleaned up dialogs, and old
coding with new functions.  Also added in more logic if case is an
ICF-MR and RTC which was not in here.  also switched the order of the
HCMI and FACI panel updates so that workers can review the faci panel.