Critical bug fix
Was overwriting current faci.  Put in logic to 1. stop the script from
overwriting the current faci. 2. stop the script if 5 FACI panels are
present (the max amount of panels), and instructs the worker to delete a
FACI panel and run the script again.  Also update MAXIS check, added
background check and reordered how the script runs to accommodate the
script ending if 5 faci panels exist.