Removing script because we forgot to add it to the beta
So, @MelissaFox-Stearns totally did this right. Unfortunately, I did not add the script to the beta MAIN MENU. So, it hasn't been tested. I'm temporarily removing it from the BETA branch so as not to confuse folks when we go live with the release today. I will add it back in later today (I hope), and this time, update the MENU to let people click on it. :)