Enhancement/bug fix
Adds a messagebox that displays expedited results in case someone runs it on an out of county case or a wrong case number. It also prevents an infinite loop in the event someone does run it on a case from another county. before it would just keep saying "navigate out of inquiry" until you hit stop script as part of function start_a_blank_case_note