Update MEMOS - SIGNIFICANT CHANGE.vbs
Thanks for your help guys! @theVKC @RobertFewins-Kalb @IlseFerris
Significant Change script includes a spec/memo denial letter that is automatically sent if you are denying the request and can put out a Tikl if client is requesting for a future month. 
The script is written with the new Do..Loop process which brings up an error message if a mandatory field is not entered.