Update NOTES - DWP BUDGET.vbs
Updated the script to make the number of DWP HHLD members a mandatory field so the personal needs calculation runs without errors. The do loop was changed to give the worker an error message if mandatory fields not entered. 
Also made slight adjustment to the spacing of the fields in the dialog box so they were no longer overlapping.