Update NAV - POLI-TEMP.vbs
I've edited it a little bit, tested, changed the dialog box so that I'm not complicating it as much as I need it to be (missing values & etc).  This should allow the user to navigate to Index and TABLE both ways; before you have to exit out back to SELF manually to use the script again.