NOTES - INCARCERATION.vbs
This is a new case noting template script that Financial workers can use to note the details of a client incarceration reported to the agency.  It will note the HH member that was incarcerated, the incarceration start date, the incarceration location, the anticipated release date if known, the facility type, how the info was received, probation officer info if known, actions taken, other notes/comments, if STAT/FACI was updated, and if a TIKL was created to check for a release date.  The script will check to see if the HH member already has 5 FACI panels and notify the Financial Worker to clean them up before proceeding.  The script has the ability to update the STAT/FACI panel for the HH member incarcerated, completing the Facility name, Facility type, FS Eligible with an N, Date In, Date Out if applicable.

The documentation for this script has already been entered on SIR in the Script Wiki.

This script resolves issue #987