Update NOTES - EVF RECEIVED.vbs
Reworked the script and dialog box to include an area for the MEMB # and the employer name. Also tweaked the dialog box to request info on how additional info was requested as "fax" isn't the only way we request info.

The script now creates a TIKL for 10 days out if additional information was requested.

I also cleaned up the "do loops" and changed it to an err_msg.

This is from issue # 1058.