Update NOTES - VERIFICATIONS NEEDED.vbs
This update fixes the script not including the case worker name in the case note.

Issue # 1158.

The TIKL part of the script is working as intended. It brings the worker to where they can enter their TIKL. The msgbox tells the worker what they are supposed to do at this point.