Fixing footer month and renewal month issue
Fixed incorrect variables, removed MAXIS_footer_finder as it seems to be working incorrectly. I will test but the function seems to be reading 11 off of MAXIS and then it asks isnumeric and it is saying it isn't. Function is likely reading in a string so the variable the function finds will never be a number and always default to CM.