Update ACTIONS - CHECK EDRS.vbs
This is an enhancement of the EDRS Checked script that runs all household members automatically. The script populates a message box with all the household members in the case and the worker can select which members to check. It then inputs the SSN and name for each member and outputs a message box detailing if there was a match for all checked. 
Thank you @CDPotter for helping me with this!!