Enabling the end user to change the Excel file
The issue was raised that the button to change wasn't doing what it advertised. I have moved the dialog building to a custom function to allow the script to loop around the dialog, resetting the scenario list, loading the new Excel file, and calling the dialog again with the new list of scenarios.

Tested and working.