Update NOTES - APPLICATION RECEIVED.vbs
pulling the MASTER version over to RELEASE. Users is SIR are reporting the footer month and year are being deleted. I switched myself to the release branch and can confirm that cases which are not in REPT/PND2 have the footer month and year removed from SELF, causing a critical script error.

When I run the script from MASTER, it be working.

Resolves #1503  -- applying @CLove update to RELEASE as this is a critical bug

BLIP: Addressing an issue reported on SIR. Cases that are not in REPT/PND2 are erroring out. Footer month and footer year are being deleted.