Update and rename NOTES - LOBBY NO SHOW.vbs to NOTES - INTERVIEW NO SHOW.vbs
This script is for both Same Day Interviews and Scheduled Interviews. If Same Day Interview is selected, the script will run what was previously known as Lobby No Show script which creates a case note when a client no shows to a lobby interview. If Scheduled Interview is selected, the script will run a dialog to gather application and interview information to enter a case note. 
Script has also been updated to incorporate the MEMO- NOMI script and gives the option to update PND2 to client delay within that script.