Fixed GA bene month bug, and HCRE
GA bene month was running through dateadd and turning something like 12/15 into 12/15/16

HCRE function is completely useless. If HC is added to TYPE MAXIS automatically navs to HCRE and enters the app info. We have no need to train on a retro case (and approval cannot handle that either) I removed the write to HCRE function. At this time I would suggest just striking through the HCRE portion on the excel spreadsheet and telling people not to fill it out rather than deleting the rows and reprogramming all of locations. The function also contained the now deleted MAXIS_dater function. Should we leave this in funclib for the time being or just delete the write to HCRE function?