fixing GA bug, and UPS variable
package approved variable needed to be reset for multiple SNAP approvals. 

MAXIS for some reason is filling in the GA in the SNAP fiat portion for 01/16 ongoing (maybe something changed in MAXIS?) added line to clear anything already written in there to FIAT the way we intended with that is stored in the array.