Fixes to NOTES >> APPROVED PROGS
BLIP: This update corrects the coordinates the script is writing on ELIG/GA and corrects the multidimensional array position location for case noting GA and MSA issuance. Previously, it was case noting the position for SNAP/MFIP Food. Finally, if the MFIP case has a significant change determination, the script gets through ELIG/MFSC to continue as normal.

Resolves #1771.