Resolves #1735: "MEMOS" is now "NOTICES"
Blip: the "MEMOS" button has been changed to a "NOTICES" button.
Agencies who update their scripts with the newest version of the
installer should see a change in the Power Pad reflecting this. The old
Power Pad will continue to work, as the MEMOS menu has been set to
forward to the NOTICES menu. All MEMOS script names have been updated to
reflect the change.