Resolved #1821- Check for expiration
Added feature to New Hire Dail Scrubber. When the script enters a Jobs panel, it checks for the "This Information will expire ..." which indicates that the JOBS panel will not carry forward to the next footer month. When this happens the script will now give you a Msgbox to alert the worker to check the next footer month.