Adding safe guard for non-matching names
BLIP: When a client gets a new hire and someone else has used their SSN the names do not match. The script was not reviewing these inconsistencies. Now the script will pause if the names are not identical and give the user a chance to review the information and stop the script if necessary.