adding cutoff date calculating
BLIP: Script will now TILK for next cutoff date to close case. Since this script is designed for the postponed wreg verification workaround we are supposed to close the case in accordance with expedited processing. This means sometimes closing cases before a ten day period to turn in verifications (to meet cutoff requirements).