Resolves #2192 ACTIONS - CASE TRANSFER memos to arep / swkr
Blip: When sending a SPEC/MEMO with new worker information, the script will now check if notices are to be sent to an AREP or Social Worker, and send copies of the memo if needed.  It also case notes the action.  A minor bug was also corrected that prevented sending the memo when transferring a privileged case.