Removed CAF rec'd portion
Since contact is made with the appt letter from the REVS scrubber, it's
appropriate to send the NOMI based on the NOMI policy. If a CAF was not
rec'd, and the appt letter was not sent, then you would not send the
NOMI, otherwise the NOMI should be sent.

This script was going against the policy. My bad on that one.