Resolves #2112 -- Creating MAGI NON-MAGI REPORT
BLIP: This script will create a report in Excel of the clients on the requested MAXIS caseloads that are active on health care. The report identifies MAGI and Non-MAGI clients. The script gives the user the option of adding information about other programs the clients are active on (CASH and SNAP). The script also gives the user the option of adding the supervisor for each caseload.