Overhaul with calculator
BLIP: Script now has better functionality regarding calculation for specific assets. 

A dynamic calculator that displays current totals for counted assets and assets applied to the BFE. When the user hits a button the calculator will refresh, it will also make sure the user doesn't enter more than $1500 into the BFE and appropriately relabel the last item to push the BFE over $1500 to BFE/counted (it will also let the user know of this change and inform them to review it). This is more in line with examples found in LTC training case notes.

A preview messagebox after the user enters all of the information. This box will display the totals, what items were selected, what their items values are, and what the status of the item is (counted excluded etc). User can then go back to change things, quit, or go to case note.

We also cleaned up some old outdated functions and case noting procedures.